requirements.txt
# Web Interface
streamlit==1.28.0

# LLM API (FREE - Groq only) - Latest stable version
groq==0.13.0

# Embeddings & Retrieval (fully compatible stack)
sentence-transformers==2.2.2
transformers==4.30.2
huggingface-hub==0.16.4
tokenizers==0.13.3
faiss-cpu==1.7.4

# Data Processing
numpy==1.24.3
pandas==2.0.0

# Visualization
plotly==5.17.0

# Environment & Config
python-dotenv==1.0.0

# Docker
gunicorn>=21.2.0

# Multimodal Support (Image Processing & CLIP)
Pillow>=9.0.0
torch>=2.0.0
open-clip-torch>=2.20.0

# PDF Export
fpdf2>=2.7.0

# Evaluation Metrics
rouge-score>=0.1.2
bert-score>=0.3.13
nltk>=3.8.0
